% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_df_names.R
\name{clean_df_names}
\alias{clean_df_names}
\title{Clean Dataframe Names}
\usage{
clean_df_names(df)
}
\arguments{
\item{df}{A Dataframe object}
}
\value{
A vector that is the names of the dataframe with spaces, periods, and
hyphens replaced with an underscore.
}
\description{
Clean Dataframe Names
}
\examples{
clean_df_names(iris)
}
