% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_names.R
\name{clean_names}
\alias{clean_names}
\title{Clean Names}
\usage{
tidy_names(df)
}
\arguments{
\item{df}{A Dataframe object}
}
\value{
A vector that is the names of the dataframe with spaces, periods, and
hyphens replaced with an underscore.
}
\description{
Clean Names
}
