% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_assumption.R
\name{lm_assumption}
\alias{lm_assumption}
\title{Linear Model Assumption Checker}
\usage{
lm_assumption(model)
}
\arguments{
\item{model}{An lm object}
}
\value{
A residual and QQ plot to check lm assumptions, as well as a VIF
calculation to measure multicollinearity.
}
\description{
Linear Model Assumption Checker
}
