% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmm_assumptions.R
\name{lmm_assumption}
\alias{lmm_assumption}
\title{Linear Mixed Models Check Assumptions}
\usage{
lmm_assumption(model)
}
\arguments{
\item{model}{A linear mixed model fitted using the lme4 package}
}
\value{
Plots for to check normality assumptions for residuals
 and random effects
}
\description{
Linear Mixed Models Check Assumptions
}
