% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refactor_cols.R
\name{refactor_cols}
\alias{refactor_cols}
\title{Refactor Columns}
\usage{
refactor_cols(df, cols)
}
\arguments{
\item{df}{A dataframe}

\item{cols}{A vector of columns that are to be turned into factors.}
}
\value{
a dataframe
}
\description{
Refactor Columns
}
