% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{scale_color_stoic}
\alias{scale_color_stoic}
\title{Scale color using AgC color palette.}
\usage{
scale_color_stoic(palette = "diverging", discrete = TRUE,
  reverse = FALSE, ...)
}
\arguments{
\item{palette:}{diverging, greens or greys}

\item{discrete:}{T or F}

\item{reverse:}{reverse the direction of the color scheme}
}
\description{
Scale color using AgC color palette.
}
