% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{scale_fill_stoic}
\alias{scale_fill_stoic}
\title{Scale fill using AgC color palette.}
\usage{
scale_fill_stoic(palette = "diverging", discrete = TRUE,
  reverse = FALSE, ...)
}
\arguments{
\item{palette:}{diverging, greens or greys}

\item{discrete:}{T or F}

\item{reverse:}{reverse the direction of the color scheme}
}
\description{
Scale fill using AgC color palette.
}
