% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_names.R
\name{tidy_names}
\alias{tidy_names}
\title{Tidy Names}
\usage{
tidy_names(df)
}
\arguments{
\item{df}{A dataframe object}
}
\value{
A vector of column names that are lowercase with spaces, periods,
and hyphens replaced with dashes
}
\description{
Tidy Names
}
