% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xkcd.R
\name{xkcd}
\alias{xkcd}
\title{View XKCD Comics}
\usage{
xkcd(comic_number = NULL, random = FALSE)
}
\arguments{
\item{comic_number}{Any number from an XKCD comic}

\item{random}{A boolean specifying a random comic}
}
\value{
An XKCD comic with relevant comic data
}
\description{
View XKCD Comics
}
\examples{
xkcd(1423) # a comic
xkcd() # most recent comic
xkcd(random = TRUE) # a random comic
\dontrun{
xkcd(961) # gif won't load in ggplot (is this possible?)
xkcd(404) # 404 page won't work
xkcd(3000) # A comic outside the current range, which fails
}
}
